requirements.txt
asgiref==3.8.1
Django==5.1.1
django-appconf==1.0.6
django-imagekit==5.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
pilkit==3.0
pillow==10.4.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
